+++ heart/config.py
@@ -20,4 +20,3 @@ class SingleDealRules:
 
 
 SINGLE = SingleDealRules()
-